+++ docs/ReleaseNotes.md
@@ -13,6 +13,7 @@ Bug Fixes:
 - Updated to remove duplicate locales in a list
 - Fixed a problem to parse xliff 2.0 files
 - Fixed a problem to merge xliff 2.0 files
+- Fixed a problem to split xliff 2.0 files
 
 Build 015
 -------